Renamed repository (#165)
* change package name

* renamed references

* updated notebooks

* update travis command

* removed remaining reference to segmentation

* pep8 from copy/paste

* reordered test functions, removed old code

* more sneaky pep8

* typo
