+++ ark/analysis/spatial_analysis_test.py
@@ -2,21 +2,19 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 import random
-from segmentation.utils import spatial_analysis
-from segmentation.utils import spatial_analysis_utils
-from segmentation.utils import synthetic_spatial_datagen
-import importlib
-importlib.reload(spatial_analysis)
+
+from ark.analysis import spatial_analysis
+from ark.utils import synthetic_spatial_datagen
 
 
-def make_threshold_mat():
+def _make_threshold_mat():
     thresh = pd.DataFrame(np.zeros((20, 2)))
     thresh.iloc[:, 1] = .5
     thresh.iloc[:, 0] = np.arange(20) + 2
     return thresh
 
 
-def make_distance_matrix(enrichment_type, dist_lim):
+def _make_distance_matrix(enrichment_type, dist_lim):
     # Make a distance matrix for no enrichment, positive enrichment, and negative enrichment
 
     if enrichment_type == "none":
@@ -59,7 +57,7 @@ def make_distance_matrix(enrichment_type, dist_lim):
         return dist_mat_neg
 
 
-def make_expression_matrix(enrichment_type):
+def _make_expression_matrix(enrichment_type):
     # Create the expression matrix with cell labels and patient labels for no enrichment,
     # positive enrichment, and negative enrichment.
 
@@ -189,11 +187,11 @@ def test_calculate_channel_spatial_enrichment():
                          "perimeter", "SampleID", "FlowSOM_ID", "cell_type"]
 
     # Test z and p values
-    marker_thresholds = make_threshold_mat()
+    marker_thresholds = _make_threshold_mat()
 
     # Positive enrichment with direct matrix initialization
-    all_data_pos = make_expression_matrix(enrichment_type="positive")
-    dist_mat_pos_direct = make_distance_matrix(enrichment_type="positive", dist_lim=dist_lim)
+    all_data_pos = _make_expression_matrix(enrichment_type="positive")
+    dist_mat_pos_direct = _make_distance_matrix(enrichment_type="positive", dist_lim=dist_lim)
 
     _, stats_pos = \
         spatial_analysis.calculate_channel_spatial_enrichment(
@@ -213,8 +211,8 @@ def test_calculate_channel_spatial_enrichment():
     assert stats_pos.loc["Point9", "z", 3, 2] > 0
 
     # Negative enrichment with direct matrix initialization
-    all_data_neg = make_expression_matrix("negative")
-    dist_mat_neg_direct = make_distance_matrix("negative", dist_lim=dist_lim)
+    all_data_neg = _make_expression_matrix("negative")
+    dist_mat_neg_direct = _make_distance_matrix("negative", dist_lim=dist_lim)
 
     _, stats_neg = \
         spatial_analysis.calculate_channel_spatial_enrichment(
@@ -234,8 +232,8 @@ def test_calculate_channel_spatial_enrichment():
     assert stats_neg.loc["Point9", "z", 3, 2] < 0
 
     # No enrichment
-    all_data_no_enrich = make_expression_matrix("none")
-    dist_mat_no_enrich = make_distance_matrix("none", dist_lim=dist_lim)
+    all_data_no_enrich = _make_expression_matrix("none")
+    dist_mat_no_enrich = _make_distance_matrix("none", dist_lim=dist_lim)
 
     _, stats_no_enrich = \
         spatial_analysis.calculate_channel_spatial_enrichment(
@@ -264,8 +262,8 @@ def test_calculate_cluster_spatial_enrichment():
                          "perimeter", "SampleID", "FlowSOM_ID", "cell_type"]
 
     # Positive enrichment with direct matrix initialization
-    all_data_pos = make_expression_matrix(enrichment_type="positive")
-    dist_mat_pos_direct = make_distance_matrix(enrichment_type="positive", dist_lim=dist_lim)
+    all_data_pos = _make_expression_matrix(enrichment_type="positive")
+    dist_mat_pos_direct = _make_distance_matrix(enrichment_type="positive", dist_lim=dist_lim)
 
     _, stats_pos = \
         spatial_analysis.calculate_cluster_spatial_enrichment(
@@ -283,8 +281,8 @@ def test_calculate_cluster_spatial_enrichment():
     assert stats_pos.loc["Point9", "z", "Pheno2", "Pheno1"] > 0
 
     # Negative enrichment with direct matrix initialization
-    all_data_neg = make_expression_matrix("negative")
-    dist_mat_neg_direct = make_distance_matrix(enrichment_type="negative", dist_lim=dist_lim)
+    all_data_neg = _make_expression_matrix("negative")
+    dist_mat_neg_direct = _make_distance_matrix(enrichment_type="negative", dist_lim=dist_lim)
 
     _, stats_neg = \
         spatial_analysis.calculate_cluster_spatial_enrichment(
@@ -302,8 +300,8 @@ def test_calculate_cluster_spatial_enrichment():
     assert stats_neg.loc["Point9", "z", "Pheno2", "Pheno1"] < 0
 
     # No enrichment
-    all_data_no_enrich = make_expression_matrix("none")
-    dist_mat_no_enrich = make_distance_matrix("none", dist_lim=dist_lim)
+    all_data_no_enrich = _make_expression_matrix("none")
+    dist_mat_no_enrich = _make_distance_matrix("none", dist_lim=dist_lim)
 
     _, stats_no_enrich = \
         spatial_analysis.calculate_cluster_spatial_enrichment(
@@ -322,8 +320,8 @@ def test_calculate_cluster_spatial_enrichment():
 
 def test_create_neighborhood_matrix():
     # get positive expression and distance matrices
-    all_data_pos = make_expression_matrix("positive")
-    dist_mat_pos = make_distance_matrix("positive", dist_lim=51)
+    all_data_pos = _make_expression_matrix("positive")
+    dist_mat_pos = _make_distance_matrix("positive", dist_lim=51)
 
     counts, freqs = spatial_analysis.create_neighborhood_matrix(all_data_pos, dist_mat_pos, distlim=51)
