+++ ark/segmentation/signal_extraction.py
@@ -1,5 +1,4 @@
 import numpy as np
-from skimage.measure import regionprops
 
 
 # TODO: work on weighted extraction and implement other discussed techniques of extraction
