+++ ark/utils/data_utils_test.py
@@ -7,12 +7,9 @@ import tempfile
 
 from mibidata import mibi_image as mi, tiff
 
-from segmentation.utils import data_utils
+from ark.utils import data_utils
 import skimage.io as io
 
-import importlib
-
-importlib.reload(data_utils)
 
 # required metadata for mibitiff writing (barf)
 METADATA = {
