+++ ark/utils/spatial_analysis_utils.py
