+++ pytest.ini
@@ -11,11 +11,6 @@ filterwarnings =
 # PEP-8 The following are ignored:
 # E501 line too long (82 > 79 characters)
 
-pep8ignore=* E501 \
-	     segmentation/training_data/* ALL \
-	     segmentation/preprocessing/* ALL \
-	     segmentation/postprocessing/* ALL \
-	     segmentation/scratch.py ALL
-
+pep8ignore=* E501
 # Enable line length testing with maximum line length of 99
-pep8maxlinelength = 99
+# pep8maxlinelength = 99
