--- segmentation/__init__.py
@@ -1,3 +1,0 @@
-from segmentation import utils
-
-from segmentation.utils import *
